+++ src/app/catalog/catalog.module.ts
@@ -1,19 +1,21 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
 
+import {SharedModule} from '../shared/shared.module';
+
 import {CatalogComponent} from './list/catalog.component';
 import {ProductComponent} from './product/product.component';
 
-import {ClickPreventDirective} from './product/click-prevent.directive';
 
 @NgModule({
     imports: [
-        CommonModule
+        CommonModule,
+
+        SharedModule
     ],
     declarations: [
         CatalogComponent,
-        ProductComponent,
-        ClickPreventDirective
+        ProductComponent
     ],
     exports: [CatalogComponent]
 })